+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 Development
 -----------
 
+* Rename HourlyBaselineData.sufficiency_warnings -> HourlyBaselineData.warnings
+* Add disqualification field to HourlyBaselineData and HourlyReportingData
 * Fix bug where HourlyBaselineData and HourlyReportingData wasn't actually NaNning zero rows when `is_electricity=True`.
 * Constrain eemeter daily model balance points to T_min_seg and T_max_seg rather than T_min and T_max.